+++ src/pipeline_telemetry/storage/memory.py
@@ -1,5 +1,6 @@
 """[summary]
 """
+import datetime
 import json
 import sqlite3
 
@@ -45,7 +46,7 @@ class TelemetryInMemoryStorage(AbstractTelemetryStorage):
         CREATE TABLE telemetry (category varchar(60),
         sub_category varchar(60), source_name varchar(40),
         process_type varchar(40), start_date_time varchar(30),
-        run_time varchar(20), telemetry_data json)"""
+        run_time varchar(20), telemetry_data json, created_at timestamp)"""
         )
 
     def store_telemetry(self, telemetry: dict) -> None:
@@ -60,7 +61,7 @@ class TelemetryInMemoryStorage(AbstractTelemetryStorage):
         json_object = json.dumps(telemetry_copy)
 
         self.db_cursor.execute(
-            "insert into telemetry values (?, ?, ?, ?, ?, ?, ?)",
+            "insert into telemetry values (?, ?, ?, ?, ?, ?, ?, ?)",
             [
                 category,
                 sub_category,
@@ -69,5 +70,6 @@ class TelemetryInMemoryStorage(AbstractTelemetryStorage):
                 start_date_time,
                 run_time_in_seconds,
                 json_object,
+                datetime.datetime.now()
             ],
         )